move periodic health check service off to a separate module/service, to separate it from HealthChecks
