+++ gwizard-healthchecks/src/main/java/com/voodoodyne/gwizard/healthchecks/HealthChecksModule.java
@@ -4,7 +4,6 @@ import com.codahale.metrics.health.HealthCheckRegistry;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
-import com.voodoodyne.gwizard.services.ServicesModule;
 import lombok.EqualsAndHashCode;
 
 @EqualsAndHashCode(of={})	// makes installation of this module idempotent
@@ -12,9 +11,6 @@ public class HealthChecksModule extends AbstractModule {
 
 	@Override
 	protected void configure() {
-		install(new ServicesModule());
-
-		bind(HealthChecksService.class).asEagerSingleton();
 	}
 
 	@Provides
