+++ gwizard-healthchecks/src/main/java/com/voodoodyne/gwizard/healthchecks/PeriodicHealthCheckConfig.java
@@ -4,7 +4,7 @@ import io.dropwizard.util.Duration;
 import lombok.Data;
 
 @Data
-public class HealthChecksConfig {
+public class PeriodicHealthCheckConfig {
 	/**
 	 * Run health checks periodically at this interval. Failing health checks will be logged at level WARN.
 	 * This accepts a flexible format like "10 minutes" or "30s"
